Optimize header syncing
Sync block headers from latest to oldest. Syncing in the random order is the artifact of
the times when the protocol used recall blocks instead of recall data chunks for proving access.

Record synced block headers - allows to get rid of traversing the block index on every startup.
Moreover, recorded block headers will make it easy to clean up the old headers to free up space
for the new headers when the disk space is limited.